# Conflicts:
#	eva/src/components/autorization/modalLog.vue
#	eva/src/components/dashBoard.vue
#	eva/src/components/dashBoardInside/dashGuntt.vue
#	eva/src/components/modalCreate.vue
#	eva/src/components/modalCreateSearch.vue
#	eva/src/components/modalDelete.vue
#	eva/src/components/modalDeleteFromMain.vue
#	eva/src/components/modalExim.vue
#	eva/src/components/modalPaper.vue
#	eva/src/components/modalReports.vue
#	eva/src/components/modalScheduler.vue
#	eva/src/components/modalSearch.vue
#	eva/src/components/modalSettings.vue
#	eva/src/components/modalThemes.vue
#	eva/src/store/store.js

+++ eva/src/components/modalThemes.vue
@@ -1,8 +1,12 @@
 <template>
-  <v-dialog
+  <modal-persistent
+    ref="modalPersistent"
     v-model="showModal"
-    persistent
-    width="400px"
+    width="400"
+    :theme="theme"
+    :persistent="isChanged"
+    :is-confirm="isChanged"
+    @cancelModal="closeModal"
   >
     <div class="themes-modal-wrapper">
       <v-card
@@ -78,7 +82,7 @@
             <div
               :style="{ color: theme.$error_color }"
               class="btn"
-              @click="deleteTheme"
+              @click="openConfirmModal"
             >
               <v-icon
                 :color="theme.$error_color"
@@ -88,6 +92,15 @@
               </v-icon>
               Удалить
             </div>
+            <modal-delete />
+            <modal-confirm
+              v-model="isConfirmModal"
+              :theme="theme"
+              :modal-text="`Вы точно хотите удалить тему - <strong>${select}</strong>?`"
+              btn-confirm-text="Удалить"
+              btn-cancel-text="Отмена"
+              @result="deleteTheme"
+            />
           </div>
           <div
             v-if="admin"
@@ -120,6 +133,7 @@
             :color="theme.$primary_button"
             outlined
             hide-details
+            @input="isChanged = true"
           />
           <div class="helper-title">
             <p @click="mode = 'manual'">
@@ -167,6 +181,7 @@
                     <v-color-picker
                       v-model="row.value"
                       dot-size="17"
+                      @input="isChanged = true"
                     />
                   </v-menu>
                 </v-col>
@@ -312,7 +327,7 @@
         </v-card-actions>
       </v-card>
     </div>
-  </v-dialog>
+  </modal-persistent>
 </template>
 
 <script>
@@ -325,14 +340,19 @@ import {
 } from '@mdi/js';
 
 export default {
+  name: 'ModalThemes',
+  model: {
+    prop: 'modalValue',
+    event: 'updateModalValue',
+  },
   props: {
-    show: {
-      type: Boolean,
-      required: true,
-    },
     admin: {
       type: Boolean,
       required: true,
+    },
+    modalValue: {
+      type: Boolean,
+      default: false,
     },
   },
   data() {
@@ -398,14 +418,76 @@ export default {
           value: '#8F8F9C',
         },
       ],
+      defaultFieldsValue: [
+        {
+          title: 'Основной фон',
+          propName: '$main_bg',
+          value: '#8F8F9C',
+        },
+        {
+          title: 'Дополнительный фон',
+          propName: '$secondary_bg',
+          value: '#8F8F9C',
+        },
+        {
+          title: 'Основные линии',
+          propName: '$main_border',
+          value: '#8F8F9C',
+        },
+        {
+          title: 'Дополнительные линии',
+          propName: '$secondary_border',
+          value: '#8F8F9C',
+        },
+        {
+          title: 'Основной текст',
+          propName: '$main_text',
+          value: '#8F8F9C',
+        },
+        {
+          title: 'Дополнительный текст',
+          propName: '$secondary_text',
+          value: '#8F8F9C',
+        },
+        {
+          title: 'Заголовки',
+          propName: '$title',
+          value: '#8F8F9C',
+        },
+        {
+          title: 'Иконки и акценты',
+          propName: '$accent_ui_color',
+          value: '#8F8F9C',
+        },
+        {
+          title: 'Взаимодейтсвия',
+          propName: '$primary_button',
+          value: '#8F8F9C',
+        },
+        {
+          title: 'Взаимодейтсвия (доп.)',
+          propName: '$primary_button_hover',
+          value: '#8F8F9C',
+        },
+      ],
       imagePreview: null,
       mode: 'select',
       error: false,
+      isChanged: false,
+      isConfirmModal: false,
     };
   },
   computed: {
-    showModal() {
-      return this.show;
+    currentThemeName() {
+      return this.$store.getters.getThemeTitle;
+    },
+    showModal: {
+      get() {
+        return this.modalValue;
+      },
+      set(value) {
+        this.$emit('updateModalValue', value);
+      },
     },
     theme() {
       return this.$store.getters.getTheme;
@@ -413,6 +495,7 @@ export default {
   },
   watch: {
     async select(selectedTheme) {
+      this.$refs.modalPersistent.focusOnModal();
       if (selectedTheme !== 'dark' && selectedTheme !== 'light') {
         const response = await fetch(`/api/theme?themeName=${selectedTheme}`);
         const themeData = await response.json();
@@ -420,9 +503,17 @@ export default {
         this.$store.commit('setTheme', content);
       } else this.$store.commit('setDefaultTheme', selectedTheme);
     },
+    showModal(val) {
+      if (!val) {
+        this.isChanged = false;
+        this.mode = 'select';
+      }
+    },
   },
   async created() {
-    await this.getThemeList();
+    await this.getThemeList().then(() => {
+      this.$set(this, 'select', this.currentThemeName);
+    });
   },
   methods: {
     resetForm() {
@@ -434,12 +525,16 @@ export default {
       });
     },
     closeModal() {
-      this.$emit('closeModal');
       this.toSelectMode();
+      this.$nextTick(() => {
+        this.showModal = false;
+      });
     },
     toSelectMode() {
       this.mode = 'select';
       this.resetForm();
+      this.isChanged = false;
+      this.$refs.modalPersistent.focusOnModal();
     },
     editTheme() {
       const themeObject = this.$store.getters.getTheme;
@@ -454,7 +549,13 @@ export default {
       this.newTitle = themeTitle;
       this.mode = 'edit';
     },
-    async deleteTheme() {
+    openConfirmModal() {
+      this.isConfirmModal = true;
+    },
+    async deleteTheme(isConfirm) {
+      if (!isConfirm) {
+        return;
+      }
       try {
         await fetch('/api/theme/delete', {
           method: 'DELETE',
@@ -469,6 +570,7 @@ export default {
       }
     },
     uploadImage() {
+      this.isChanged = true;
       const image = this.$refs.imageInput.files[0];
       const reader = new FileReader();
       reader.readAsDataURL(image);
@@ -516,8 +618,12 @@ export default {
           const themeData = await res.json();
           const content = JSON.parse(themeData);
           this.$store.commit('setTheme', content);
+          this.$set(this, 'select', this.currentThemeName);
           this.mode = 'select';
-          await this.getThemeList();
+          await this.getThemeList().then(() => {
+            this.isChanged = false;
+            this.$refs.modalPersistent.focusOnModal();
+          });
         } catch (e) {
           console.log(e);
         }
